Fix stack misalignment issue on saved thread state
When saving the thread state, the code did not account for the five registers (eip, cs, eflags, sp, ss) that the CPU automatically pushes to the stack as part of the interrupt. Since the registers are pushed to the stack before the pushad in `isr_common_stub`, the saved esp value is incorrect and offset by 20 bytes. This is why each thread context switch resulted in one thread's stack position being shifted by -0x14. The ultimate consequence of this was a crash after the thread stacks grew downwards far enough to interfere with the kernel stack.

The fix is simple: we can just add 20 to the saved esp value in `save_thread_state` to compensate for this offset.